Renamed Infrastructure to Shared to match other assemblies
