+++ src/AdminAssistant.Blazor/Server/WebAPI/v1/AccountsModule/BankAccountInfoController.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using AdminAssistant.DomainModel.Infrastructure;
+using AdminAssistant.DomainModel.Shared;
 using AdminAssistant.DomainModel.Modules.AccountsModule.CQRS;
 using AdminAssistant.Framework.Providers;
 using AutoMapper;
