+++ src/AdminAssistant.Infra/DAL/EntityFramework/Model/CoreSchema.cs
@@ -1,4 +1,4 @@
-using AdminAssistant.DomainModel.Infrastructure;
+using AdminAssistant.DomainModel.Shared;
 using AdminAssistant.DomainModel.Modules.AccountsModule;
 using Microsoft.EntityFrameworkCore;
