+++ src/AdminAssistant.Test/WebAPI/v1/AccountsModule/BankAccountInfoController_UnitTest.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using AdminAssistant.DomainModel;
-using AdminAssistant.DomainModel.Infrastructure;
+using AdminAssistant.DomainModel.Shared;
 using AdminAssistant.DomainModel.Modules.AccountsModule;
 using AdminAssistant.DomainModel.Modules.AccountsModule.CQRS;
 using Ardalis.Result;
