+++ src/AdminAssistant/DomainModel/Shared/ConfigurationSettings.cs
@@ -1,4 +1,4 @@
-namespace AdminAssistant.DomainModel.Infrastructure
+namespace AdminAssistant.DomainModel.Shared
 {
     public class ConfigurationSettings
     {
