+++ src/AdminAssistant/DomainModel/Shared/UserContextProvider.cs
@@ -1,4 +1,4 @@
-namespace AdminAssistant.DomainModel.Infrastructure
+namespace AdminAssistant.DomainModel.Shared
 {
     public interface IUserContextProvider
     {
